requirements.txt
transformers
torch
torchvision
pillow
numpy
requests
python-dotenv
gradio
replicate